refac: Segregate data product stuff in settlement reports job
