+++ source/databricks/settlement_report/settlement_report_job/domain/settlement_report_args.py
@@ -15,7 +15,7 @@ from uuid import UUID
 from dataclasses import dataclass
 from datetime import datetime
 
-from settlement_report_job.domain.calculation_type import CalculationType
+from settlement_report_job.infrastructure.calculation_type import CalculationType
 from settlement_report_job.domain.market_role import MarketRole
