+++ source/databricks/settlement_report/settlement_report_job/domain/time_series/time_series_factory.py
@@ -17,9 +17,6 @@ from uuid import UUID
 from pyspark.sql import DataFrame
 
 from settlement_report_job.domain.market_role import MarketRole
-from settlement_report_job.domain.DataProductValues.metering_point_resolution import (
-    MeteringPointResolutionDataProductValue,
-)
 from settlement_report_job.domain.repository import WholesaleRepository
 from settlement_report_job.domain.time_series.prepare_for_csv import (
     prepare_for_csv,
@@ -29,6 +26,9 @@ from settlement_report_job.domain.time_series.read_and_filter import (
 )
 from settlement_report_job.logger import Logger
 from settlement_report_job.infrastructure import logging_configuration
+from settlement_report_job.wholesale.data_values import (
+    MeteringPointResolutionDataProductValue,
+)
 
 log = Logger(__name__)
