+++ source/databricks/settlement_report/settlement_report_job/infrastructure/calculation_type.py
@@ -16,6 +16,8 @@ from enum import Enum
 
 
 class CalculationType(Enum):
+    """The job parameter values must correspond with these values."""
+
     BALANCE_FIXING = "balance_fixing"
     WHOLESALE_FIXING = "wholesale_fixing"
     FIRST_CORRECTION_SETTLEMENT = "first_correction_settlement"
