+++ source/databricks/settlement_report/settlement_report_job/infrastructure/paths.py
@@ -1,0 +1,2 @@
+def get_settlement_reports_output_path(catalog_name: str) -> str:
+    return f"/Volumes/{catalog_name}/wholesale_settlement_report_output/settlement_reports"  # noqa: E501
