+++ source/databricks/settlement_report/settlement_report_job/wholesale/__init__.py
@@ -1,0 +1,6 @@
+"""
+This sub-package contains the schemas, column names, and more from wholesale data products
+on which the settlement report job depends.
+
+See the data products at https://energinet.atlassian.net/wiki/spaces/D3/pages/849903618/Wholesale+Data+Products.
+"""
