+++ source/databricks/settlement_report/settlement_report_job/wholesale/database_definitions.py
@@ -23,7 +23,3 @@ class WholesaleBasisDataDatabase:
 class WholesaleWholesaleResultsDatabase:
     DATABASE_NAME = "wholesale_results"
     ENERGY_V1_VIEW_NAME = "energy_v1"
-
-
-def get_settlement_reports_output_path(catalog_name: str) -> str:
-    return f"/Volumes/{catalog_name}/wholesale_settlement_report_output/settlement_reports"  # noqa: E501
