+++ source/databricks/settlement_report/settlement_report_job/wholesale/schemas/charge_price_information_periods_v1.py
