+++ source/databricks/settlement_report/settlement_report_job/wholesale/schemas/metering_point_time_series_v1.py
