+++ source/databricks/settlement_report/tests/conftest.py
@@ -18,7 +18,7 @@ from typing import Callable, Generator
 
 from delta import configure_spark_with_delta_pip
 from pyspark.sql import SparkSession
-from settlement_report_job.domain.calculation_type import CalculationType
+from settlement_report_job.infrastructure.calculation_type import CalculationType
 from settlement_report_job.domain.market_role import MarketRole
 from settlement_report_job.domain.settlement_report_args import SettlementReportArgs
 from tests.fixtures import DBUtilsFixture
