+++ source/databricks/settlement_report/tests/domain/test_csv_writer.py
@@ -18,18 +18,16 @@ import pyspark.sql.functions as F
 from tests.fixtures import DBUtilsFixture
 from functools import reduce
 import pytest
-from settlement_report_job.domain.DataProductValues.metering_point_type import (
-    MeteringPointTypeDataProductValue,
-)
-from settlement_report_job.domain.DataProductValues.metering_point_resolution import (
-    MeteringPointResolutionDataProductValue,
-)
 from settlement_report_job.domain.report_data_type import ReportDataType
 
 from settlement_report_job.domain.settlement_report_args import SettlementReportArgs
 import tests.test_factories.time_series_csv_factory as factory
-from settlement_report_job.infrastructure.column_names import (
+from settlement_report_job.infrastructure.csv_column_names import (
     TimeSeriesPointCsvColumnNames,
+)
+from settlement_report_job.wholesale.data_values import (
+    MeteringPointResolutionDataProductValue,
+    MeteringPointTypeDataProductValue,
 )
