+++ source/databricks/settlement_report/tests/domain/test_report_name_factory.py
@@ -4,7 +4,7 @@ from datetime import datetime
 import pytest
 from pyspark.sql import SparkSession
 
-from settlement_report_job.domain.calculation_type import CalculationType
+from settlement_report_job.infrastructure.calculation_type import CalculationType
 from settlement_report_job.domain.market_role import MarketRole
 from settlement_report_job.domain.report_name_factory import (
     FileNameFactory,
