+++ source/databricks/settlement_report/tests/domain/time_series/test_read_and_filter.py
@@ -12,13 +12,13 @@ import test_factories.charge_price_information_periods_factory as charge_price_i
 
 
 from settlement_report_job.domain.market_role import MarketRole
-from settlement_report_job.domain.DataProductValues.metering_point_resolution import (
-    MeteringPointResolutionDataProductValue,
-)
 from settlement_report_job.domain.time_series.read_and_filter import (
     read_and_filter_for_wholesale,
 )
-from settlement_report_job.infrastructure.column_names import DataProductColumnNames
+from settlement_report_job.wholesale.column_names import DataProductColumnNames
+from settlement_report_job.wholesale.data_values import (
+    MeteringPointResolutionDataProductValue,
+)
 
 DEFAULT_FROM_DATE = default_data.DEFAULT_FROM_DATE
 DEFAULT_TO_DATE = default_data.DEFAULT_TO_DATE
